server socket functionality
